Remove dead platform code
The Task struct and Platform::pump_message_loop() method were
non-functional, remove them.
