+++ src/lib.rs
@@ -111,10 +111,6 @@ pub use module::*;
 pub use object::*;
 pub use platform::new_default_platform;
 pub use platform::Platform;
-pub use platform::Task;
-// TODO(ry) TaskBase and TaskImpl ideally shouldn't be part of the public API.
-pub use platform::TaskBase;
-pub use platform::TaskImpl;
 pub use primitives::*;
 pub use private::*;
 pub use promise::{PromiseRejectEvent, PromiseRejectMessage, PromiseState};
